// src/types/types.ts
export type NetworkType = "local" | "testnet" | "mainnet" | "futurenet"

export type Network = {
	id: NetworkType
	label: string
	horizonUrl: string
	horizonHeaderName?: string
	horizonHeaderValue?: string
	rpcUrl: string
	rpcHeaderName?: string
	rpcHeaderValue?: string
	passphrase: string
}
// ...
export type SignTransactionFn = (
	xdr: string,
	opts?: {
		networkPassphrase?: string
		address?: string
		path?: string
		submit?: boolean
		submitUrl?: string
	},
) => Promise<{
	signedTxXdr: string
	signerAddress?: string
}>